refactor(data): reorganize pronunciation and location data files
- Move Pronunciation.tsx and StLucia.tsx files to new Misc directory
- Update import paths in affected components
- Fix Wikipedia link label for Saint Lucian Creole
